Seems like this usage reduced the number of "peaks".

With the old code:
It showed an average of 3.2-2.1ms per call
And there were occasional "peaks" up to 4.5

With the new code:
It shows an average of 2.9-2.1ms per call
And very rare "peaks" up to ~3.5.
![изображение](https://user-images.githubusercontent.com/739505/222908445-fbe1190e-01e2-436c-abec-6b3bb021d6ab.png)