@Antizombie 2023-03-04
I think I will also try to use "goto" to immediately bypass part of the code. For example, when checking 

> local blockedForward = (linkParams.Direction == "Forward" and link.Nodes[2] == node)

it's possible to jump directly to the next node (node.LinkByLinkedNode) without going through the rest of the code.